requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
minio==7.2.8
ultralytics==8.2.100
easyocr==1.7.1
opencv-python==4.10.0.84
pillow==10.4.0
pyjwt==2.9.0
flask-mail==0.10.0
textdistance==4.6.3
Bcrypt-Flask==1.0.2
gunicorn==23.0.0
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
cryptography==43.0.1